+++ packages/theme/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [3.11.0](https://github.com/p3ol/junipero/compare/v3.10.7...v3.11.0) (2025-07-30)
+
+
+### Features
+
+* **react:** add infinite canvas ([48fccca](https://github.com/p3ol/junipero/commit/48fccca81c040a92b15c5411287272908178395b))
+* require node 20 min ([6eaa695](https://github.com/p3ol/junipero/commit/6eaa6959c6b28dd27d24b00d79e0ff376e3005ef))
+
+
+
+
+
 ## [3.10.4](https://github.com/p3ol/junipero/compare/v3.10.3...v3.10.4) (2025-02-03)
 
 **Note:** Version bump only for package @junipero/theme